Added new accepted command to allow grouped rolls, Fixed request and reply structure access to markdown reply
